Solved problem 96
The solution could be improved because the solution currently takes more
than a minute to solve.  Currently the solution only uses a guess and test
method.  The only optimization is that any squares for which only one
number is possible are automatically filled in.

I also converted the solutions for problems 4, 5, and 15 from a Mathematica
notebook to a Mathematica script.